Initial implementation of wave summation
Currently it's written to calculate the first 10 values from rangeStart
mostly as a sanity check. Next up I'd like to test with a 1000 or so
values and plot their graph to see what the pattern is like. After that,
I'll run the program for the entire range we decided on.